auto_generation_pytest/pyproject.toml
[tool.poetry]
name = "auto_generation_pytest"
version = "0.1.6"
description = "auto_generation_pytest"
authors = ["moku27 <xieshi@forchange.tech>"]

[tool.poetry.dependencies]
python = "^3.7"
allpairspy = "^2.5.0"
grpcio = "^1.30.0"
pytest = "^5.4.3"
allure-pytest = "^2.8.16"
grpcio-tools = "^1.30.0"
python-dotenv = "^0.14.0"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
